### 2025-09-05 #8937 @editorialbot
[PRE REVIEW]: HLAfreq: Download and combine HLA allele frequency data
**Submitting author:** @DAWells (<a href="http://orcid.org/0000-0002-4531-5968">David A. Wells</a>)
**Repository:** https://github.com/BarinthusBio/HLAfreq
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.0.5
**Managing EiC:** Rachel Kurchin
