### 2025-09-12 3305884
Add doi's where available
