+++ paper/paper.bib
@@ -45,6 +45,7 @@ MRREVIEWER = {Michael J. Evans},
   title={HLA allele selection for designing peptide vaccines},
   author={Gulukota, Kamalakar and DeLisi, Charles},
   journal={Genetic Analysis: Biomolecular Engineering},
+  doi={10.1016/1050-3862(95)00156-5},
   volume={13},
   number={3},
   pages={81--86},
@@ -63,6 +64,7 @@ MRREVIEWER = {Michael J. Evans},
   title={The HLA region and autoimmune disease: associations and mechanisms of action},
   author={Simmonds, MJ and Gough, SCL},
   journal={Current genomics},
+  doi={10.2174/138920207783591690},
   volume={8},
   number={7},
   pages={453--465},
@@ -73,6 +75,7 @@ MRREVIEWER = {Michael J. Evans},
   title={The clinical significance of human leukocyte antigen (HLA) allele compatibility in patients receiving a marrow transplant from serologically HLA-A, HLA-B, and HLA-DR matched unrelated donors},
   author={Morishima, Yasuo and Sasazuki, Takehiko and Inoko, Hidetoshi and Juji, Takeo and Akaza, Tatsuya and Yamamoto, Ken and Ishikawa, Yoshihide and Kato, Shunichi and Sao, Hiroshi and Sakamaki, Hisashi and others},
   journal={Blood, The Journal of the American Society of Hematology},
+  doi={10.1182/blood.V99.11.4200},
   volume={99},
   number={11},
   pages={4200--4206},
@@ -83,6 +86,7 @@ MRREVIEWER = {Michael J. Evans},
   title={HLA matching in unrelated stem cell transplantation up to date},
   author={F{\"u}rst, Daniel and Neuchel, Christine and Tsamadou, Chrysanthi and Schrezenmeier, Hubert and Mytilineos, Joannis},
   journal={Transfusion Medicine and Hemotherapy},
+  doi={10.1159/000502263},
   volume={46},
   number={5},
   pages={326--336},
