### 2025-09-12 28e374b
fix affiliation for McAuley
