+++ paper/paper.md
@@ -24,14 +24,12 @@ bibliography: paper.bib
 ---
 
 # Summary
-Human leukocyte antigen (HLA) genes encode cell-surface proteins which play an important role in immunity. Since different HLA alleles enable different immune responses, the population frequency of HLA alleles is often considered when designing vaccines [@gulukota1996hla]. Specific HLA alleles have been linked to autoimmune disease [@simmonds2007hla] and associated with adverse drug reactions [@fan2017hla]. Further, the success of solid organ and stem cell transplants is related to HLA matching between donor and recipient [@morishima2002clinical; furst2019hla].
+Human leukocyte antigen (HLA) genes encode cell-surface proteins which play an important role in immunity. Since different HLA alleles enable different immune responses, the population frequency of HLA alleles is often considered when designing vaccines [@gulukota1996hla]. Specific HLA alleles have been linked to autoimmune disease [@simmonds2007hla] and associated with adverse drug reactions [@fan2017hla]. Further, the success of solid organ and stem cell transplants is related to HLA matching between donor and recipient [@morishima2002clinical; @furst2019hla].
 
 The [Allele Frequency Net Database](www.allelefrequencies.net) is a publicly available repository for human immune gene frequency data from across the world [@Gonzalez-Galarza2020]. However, difficulties downloading and combining data from multiple studies make it hard for researchers to study larger regions or even single countries where the data is split across many sources. To address this gap, we present `HLAfreq`: a Python package which can be used to download, combine and analyse datasets from the Allele Frequency Net Database.
 
 # Statement of need
-The Allele Frequency Net Database is an excellent resource; however, downloading data from a large number of studies is currently manual and slow. After downloading multiple studies, combining them is hindered by different allele resolutions, missing alleles, and incomplete studies. `HLAfreq` provides functions to identify incomplete studies, handle missing alleles, harmonise allele resolution, calculate population coverage, and estimate allele frequencies and uncertainty using a Bayesian framework. Allele frequency plots can be generated to identify anomalous datasets and interesting diversity in a set of populations. To get started, see the guide and examples at [github.com/BarinthusBio/HLAfreq](https://github.com/BarinthusBio/HLAfreq).
-
-The function `combineAF()` is used to combine multiple datasets and estimate the allele frequencies according to the 'default model' described below. So that larger studies contribute more to the combined allele frequency estimate, each dataset is weighted by twice the sample size by default (because each individual is diploid). Alternatively, any supplied weighting can be used; for example, population size can be used for a multi-country estimate, see the [multi-country example](https://barinthusbio.github.io/HLAfreq/HLAfreq/examples/multi_country.html). `combineAF()` performs several automated checks on the dataset. Studies are flagged as incomplete if the total allele frequency is outside a specified range (0.95-1.1 by default). Unmeasured alleles are added with a frequency of zero to ensure all populations report the same set of alleles before combining. Allele resolution is automatically checked because only alleles of the same resolution can be combined. The allele frequencies and credible intervals can be estimated according to the 'compound model' described below with `AFhdi()` imported from the submodule `HLAfreq.HLAfreq_pymc`.
[…]
