### 2025-09-13 24a0533
Shorten statistical methods section
