+++ paper/paper.md
@@ -34,10 +34,10 @@ The Allele Frequency Net Database is an excellent resource; however, downloading
 # Methods
 
 ## Statistical methods
-`HLAfreq` uses a Bayesian framework to estimate allele frequency statistics from combined datasets for a specific population. The user can select from two possible statistical models. The simpler 'default model' gives point estimates for allele frequencies. The more sophisticated 'compound model' gives both point estimates and credible intervals.
+`HLAfreq` uses a Bayesian framework to estimate allele frequency statistics from combined datasets for a specific population. The user can select from two  statistical models. The simpler 'default model' gives point estimates for allele frequencies. The more sophisticated 'compound model' gives both point estimates and credible intervals.
 
 ### Default model
-Let $p_k$ be the frequency of the $k$-th allele of a particular gene in a given population (e.g. a country). The default model assumes that the observations from all datasets for the population are drawn independently and that the probability of being the $k$-th allele is $p_k$. In other words, each observation is drawn from a categorical distribution with parameters $(p_1,\dots,p_K)$ where $K$ is the total number of alleles. The prior for $(p_1,\dots,p_K)$ is taken to be a Dirichlet distribution with parameters $\alpha_1,\dots,\alpha_K$ which are chosen by the user. The Dirichlet distribution is a generalisation of the Beta distribution to higher dimensions; its properties are described in Section 4.6.3 of [@murphy2022probabilistic].
+Let $p_k$ be the frequency of the $k$-th allele of a particular gene in a given population (e.g. a country). The default model assumes that the observations from all datasets for the population are drawn independently and that the probability of being the $k$-th allele is $p_k$. In other words, each observation is drawn from a categorical distribution with parameters $(p_1,\dots,p_K)$ where $K$ is the total number of alleles. The prior for $(p_1,\dots,p_K)$ is taken to be a Dirichlet distribution with parameters $\alpha_1,\dots,\alpha_K$. The Dirichlet distribution is a generalisation of the Beta distribution to higher dimensions; see Section 4.6.3 of [@murphy2022probabilistic].
 
 The Dirichlet distribution is conjugate to the categorical distribution, meaning that the posterior distribution for the default model is also Dirichlet. More precisely, if the combined datasets contain $x_k$ observations of the $k$-th allele (for $k=1,\dots,K$) then the posterior distribution is Dirichlet with parameters $\alpha_1+x_1,\dots,\alpha_K+x_K$. The posterior mean for the frequency of allele $j$ is then given by
 $$
@@ -46,10 +46,10 @@ $$
 
 By default, `HLAfreq` takes the prior parameters to be $\alpha_1=\dots=\alpha_K=1$. This results in a uniform prior on $(p_1,\dots,p_K)$ subject to the constraints that $p_1,\dots,p_K\geq 0$ and $p_1+\dots+p_K=1$. The user can specify alternative values for $\alpha_1,\dots,\alpha_K$. These parameters may be interpreted as a `pseudocount' in the sense that choosing the prior $\alpha_1,\dots,\alpha_K$ is equivalent to taking a uniform prior and then observing a dataset with $\alpha_k-1$ observations of the $k$-th allele. (Intuitively the uniform prior corresponds to one observation of each allele). This can be used as a heuristic for choosing prior parameters based on external information.
 
-The default model could in principle be used to estimate credible intervals for the frequencies $p_1,\dots,p_K$ however this option is not provided by `HLAfreq` because in practice such intervals are frequently unrealistically narrow. This is because the default model does not account for variance between studies. In the next subsection we outline how we can account for this variation and obtain accurate credible intervals. The current model is chosen as the default because it is simpler and we expect its point estimates to be sufficient for the majority of use cases.
[…]
