### 2025-09-15 #8937 @Kevin-Mattheus-Moerman
@DAWells can you work on the above points ☝️

Please note also that JOSS is currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy. For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!
