### 2025-09-15 b57b8bd
Add contribution and community guidlines
