+++ CONTRIBUTING.md
@@ -1,0 +1,36 @@
+# Community guidelines
+Thank you for using `HLAfreq`!
+It's great that you'd like to contribute in any form!
+Asking questions as issues is an important contribution.
+See below for particular information to include. If you see an
+issue you can help with, your answer would be greatly appreciated.
+When creating or answering issues please be kind!
+
+## Seek support
+If you need help using this tool, the first port of call is the
+examples and documentation [here](https://barinthusbio.github.io/HLAfreq/HLAfreq.html).
+But if the answer is not there, please open an [issue](https://github.com/BarinthusBio/HLAfreq/issues).
+
+## Report problems with the software
+If the software isn't working or you think you've found a bug,
+please open an [issue](https://github.com/BarinthusBio/HLAfreq/issues). Please include:
+- the version of `HLAfreq`
+- code to reproduce the error
+- any error message/the result
+- what you expected/think should have happened instead
+
+## Contribute to the software
+If you want to add to `epitope_aligner` that's great! Fork the
+repo and create a branch named after the change you're making.
+Make your contribution to the new branch on your fork and add
+tests to verify that your code runs as expected. The tests should
+pass when run by nox before submitting a pull request. See this
+[guide](https://learn.scientific-python.org/development/guides/tasks/)
+on using nox.
+
+When you're happy with your changes and the tests are passing,
+submit a pull request. In the pull request describe the motivation
+for your changes, their impact, and the tests you have written for
+them.
+
+# Thank you for contributing!
