+++ README.md
@@ -134,6 +134,13 @@ documentation in `./docs`.
 <!-- Documentation generated by pdoc should not be commited
 as it is auto generated by a github action. -->
 
+## Community guidelines
+Thank you for using `HLAfreq`! Contributions in any form
+are immensely helpful: asking questions, answering issues,
+and pull requests are all great.
+
+For full details see [CONTRIBUTING.md](https://github.com/BarinthusBio/HLAfreq/blob/main/examples/quickstart.py). In short, if the answer to
+your question is not in the docs, open an [issue](https://github.com/BarinthusBio/HLAfreq/issues). If you'd like to improve `HLAfreq` create a fork and a pull request.
 
 <!-- ## Developer notes
 Install in dev mode
