### 2025-09-15 #8937 @DAWells
Thanks @rkurchin and @Kevin-Mattheus-Moerman
- I have shortened the paper
- added the missing affiliation
- added the missing doi's
- added `CONTRIBUTING.md`
- referenced `CONTRIBUTING.md` in the `README.md`

## Suggested reviewers
- darnells
- vegetableyu
- sdwfrost
- mastoffel
- tkchafin
