### 2026-01-09 c090571
AddResearch Impact Statement, Software Design, and AI usage disclosure sections
Shorten the main text to account for these new sections
