### 2026-01-09 a19c29d
AddResearch Impact Statement, Software Design, and AI usage disclosure sections
Shorten the main text to account for these new sections
