+++ paper/paper.md
@@ -26,15 +26,15 @@ bibliography: paper.bib
 # Summary
 Human leukocyte antigen (HLA) genes encode cell-surface proteins which play an important role in immunity. Since different HLA alleles enable different immune responses, the population frequency of HLA alleles is often considered when designing vaccines [@gulukota1996hla]. Specific HLA alleles have been linked to autoimmune disease [@simmonds2007hla] and associated with adverse drug reactions [@fan2017hla]. Further, the success of solid organ and stem cell transplants is related to HLA matching between donor and recipient [@morishima2002clinical; @furst2019hla].
 
-The [Allele Frequency Net Database](www.allelefrequencies.net) is a publicly available repository for human immune gene frequency data from across the world [@Gonzalez-Galarza2020]. However, difficulties downloading and combining data from multiple studies make it hard for researchers to study larger regions or even single countries where the data is split across many sources. To address this gap, we present `HLAfreq`: a Python package which can be used to download, combine and analyse datasets from the Allele Frequency Net Database.
+We present `HLAfreq`: a Python package which can be used to download, combine and analyse multiple HLA allele frequency datasets.
 
 # Statement of need
-The Allele Frequency Net Database is an excellent resource; however, downloading data from a large number of studies is currently manual and slow. After downloading multiple studies, combining them is hindered by different allele resolutions, missing alleles, and incomplete studies. `HLAfreq` provides functions to identify incomplete studies, handle missing alleles, harmonise allele resolution, calculate population coverage, and estimate allele frequencies and uncertainty using a Bayesian framework. When combining studies, estimates are weighted by twice the sample size (because each individual is diploid). Alternatively, any supplied weighting can be used, see the [multi-country example](https://barinthusbio.github.io/HLAfreq/HLAfreq/examples/multi_country.html). Allele frequency plots can be generated to identify anomalous datasets and interesting diversity in a set of populations. To get started, see the guide and examples at [github.com/BarinthusBio/HLAfreq](https://github.com/BarinthusBio/HLAfreq).
+The [Allele Frequency Net Database](www.allelefrequencies.net) is a publicly available repository for human immune gene frequency data from across the world [@Gonzalez-Galarza2020]. However, downloading data from a large number of studies is currently manual and slow. After downloading multiple studies, combining them is hindered by different allele resolutions, missing alleles, and incomplete studies. `HLAfreq` provides functions to identify incomplete studies, handle missing alleles, harmonise allele resolution, calculate population coverage, and estimate allele frequencies and uncertainty using a Bayesian framework. Allele frequency plots can be generated to identify anomalous datasets and interesting diversity in a set of populations. To get started, see the guide and examples at [github.com/BarinthusBio/HLAfreq](https://github.com/BarinthusBio/HLAfreq).
 
 # Methods
 
 ## Statistical methods
-`HLAfreq` uses a Bayesian framework to estimate allele frequency statistics from combined datasets for a specific population. The user can select from two  statistical models. The simpler 'default model' gives point estimates for allele frequencies. The more sophisticated 'compound model' gives both point estimates and credible intervals.
+`HLAfreq` uses a Bayesian framework to estimate allele frequency statistics from combined datasets for a specific population. The user can select from two statistical models. The simpler 'default model' gives point estimates for allele frequencies. The more sophisticated 'compound model' gives both point estimates and credible intervals.
 
 ### Default model
[…]
