+++ noxfile.py
@@ -13,7 +13,7 @@ def bandit(session):
     session.install("bandit")
     session.run("bandit", "-r", "src")
 
-@nox.session(python=["3.10", "3.11", "3.12"])
+@nox.session(python=["3.10", "3.11", "3.12", "3.13", "3.14"])
 def tests(session):
     session.install("pytest", "pytest-cov")
     session.install(".")
