### 2026-02-14 #8937 @fboehm
@Yegor13 @roman-tremmel @juliofdiaz @sdwfrost @mastoffel @dwinter - would you like to review this submission for JOSS? Our reviews are checklist-driven. Thanks for considering this invitation.
