### 2026-02-19 #8937 @fboehm
@greggj2016 @matthewfeickert @assignUser @jdrugo @harveyaa - would you like to review this submission for JOSS?

Thanks for considering the invitation!
