### 2026-02-19 #8937 @matthewfeickert
> would you like to review this submission for JOSS?

@fboehm while I would like to review a JOSS paper again in the future, I am currently already fully booked up with my JOSS editor responsibilities, so I'm going to have to pass. Thanks for asking though. 👍
