### 2026-02-19 #8937 @jdrugo
> [@greggj2016](https://github.com/greggj2016) [@matthewfeickert](https://github.com/matthewfeickert) [@assignUser](https://github.com/assignUser) [@jdrugo](https://github.com/jdrugo) [@harveyaa](https://github.com/harveyaa) - would you like to review this submission for JOSS?
>
> Thanks for considering the invitation!

@fboehm I am not a subject level expert for this submission. Thus, I would not be able to assess whether the submitted software functions correctly. Furthermore, I have not participated in a JOSS review before, and so would need some hand-holding to get started. If neither is an issue, I'd be happy to review this submission.
