### 2026-02-20 #8937 @rkurchin
> [@fboehm](https://github.com/fboehm) I am not a subject level expert for this submission. Thus, I would not be able to assess whether the submitted software functions correctly. Furthermore, I have not participated in a JOSS review before, and so would need some hand-holding to get started. If neither is an issue, I'd be happy to review this submission.

@fboehm, I'm concerned that this reviewer may not be a good fit for this submission. Given that they explicitly say they can't assess the software, I suggest you look for someone else.
