### 2026-02-22 #8937 @usethedata
@fboehm -- Yes, I can review this.  I spent a bit of time going over the repo, associated documentation, and the source paper.  I'm comfortable that this is close enough to my education and work I've done for me to assess.

@jdrugo -- I've done multiple reviews and was a topic editor for a while.  Whatever you wind up doing in your first reviews (whether this paper or something else), please feel free to reach out and I will be happy to coach you through the process.
