### 2026-02-23 #8937 @usethedata
@DAWells -- A couple of minor items before the actual review issue gets opened up:

1) In CONTRIBUTING.md, under Contribute to the software, there is a reference to epitope_aligner, rather than HLAfreq.

2) In README.md, under Community Guidelines, the link to CONTRIBUTING.md actually points to quickstart.py in the examples folder.
