### 2026-02-24 #8937 @fboehm
@darnells @vegetableyu @tkchafin - would you like to review this submission for JOSS? our reviews are checklist-driven. You can find an example checklist here: https://github.com/openjournals/joss-reviews/issues/9891#issuecomment-3936842337

Thanks for considering this invitation.
