### 2026-02-26 #8937 @fboehm
@greggj2016 @trangdata - would you like to review this submission for JOSS? Our reviews are checklist-driven. You can find an example checklist here: https://github.com/openjournals/joss-reviews/issues/9891#issuecomment-3936842337

Thank you for considering this invitation.
