### 2026-02-27 #10122 @editorialbot
[REVIEW]: HLAfreq: Download and combine HLA allele frequency data
**Submitting author:** @DAWells (<a href="http://orcid.org/0000-0002-4531-5968">David A. Wells</a>)
**Repository:** https://github.com/BarinthusBio/HLAfreq
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.6
