### 2026-03-08 #10122 @jdrugo
@fboehm Given that this is my first JOSS review, I would need some guidelines on the following checklist items:
* **Contribution and authorship**: Are all JOSS paper authors expected to have contributed to the software? Only the first author seems to have contributed.
* **Collaborative effort**: The software appears to have a single contributing author with no community input so far. Therefore, I cannot check this item. Should all items be checked in order to complete the review?

Thanks!
