### 2026-03-08 #10122 @fboehm
> [@fboehm](https://github.com/fboehm) Given that this is my first JOSS review, I would need some guidelines on the following checklist items:
>
> * **Contribution and authorship**: Are all JOSS paper authors expected to have contributed to the software? Only the first author seems to have contributed.
> * **Collaborative effort**: The software appears to have a single contributing author with no community input so far. Therefore, I cannot check this item. Should all items be checked in order to complete the review?
>
> Thanks!

Thank you, @jdrugo, for the questions.

For the first point, it is not necessary that every author has authored a git commit. There are many other ways for an author to contribute to the project. For example, they may have planned the project.

For the second point, It's ok for only one person to have authored git commits. I think that by posting the repository in a public setting like on github, this criterion is satisfied.

And yes, all boxes must be checked before the publication can occur. Typically, a reviewer may not check all boxes on initial review. Instead, the reviewer will point out the revisions that are needed to satisfy a given criterion.

@rkurchin - do you have anything to add to these reponses?
