### 2026-03-08 #10122 @usethedata
This review was conducted based on commit `f753c9b9` in the `main` branch and `c0905714` in the paper branch.

Overall, I recommend this for publication, provided that a sufficient **State of the field** section is added to the paper.  While I didn't list it as a blocker, I do think the pandas 3 compatibility issue should also be addressed.

I am flagging the *Contribution and authorship* and *Research impact statement* for @fboehm editorial comment.
  - I think the Research impact statement is OK. Having worked in industry, I can see where this may be difficult to expand.  It doesn't hit the "Not Acceptable" criteria for me, so I'm good with passing it here.  But I'd also ask the authors if they see a way to strengthen this.
  - For *Contribution and authorship", I think some clear statement of McAuley's contribution is needed.  I remember reading this at some point, but I'm not seeing it in the paper or other materials.

## Items that block acceptance of this paper:
- [ ] **State of the field:** As noted by editorialbot, the paper does not contain a "State of the field" section. This is a required element. The Statement of Need alludes to the absence of automation in AFND access but does not compare HLAfreq to any alternative approaches.

## Items that should be addressed, but which I do not see as blockers:
- [~] **Automated tests:** There is a suite of 64 tests, which passed for me on Python 3.14.3.  GitHub CI is configured to run these tests on pushes to `main` and `dev`.  However, `test_correct_c_array` and `test_correct_c_array_alleles` lack `assert` statements (`test_HLAfreq_pymc.py` lines 31 and 40), so I believe these "tests" will always pass.
- **Pandas 3 incompatibility:** In running the sample notebooks, I ran into what I believe to be a compatibility issue with Pandas 3.  Pandas 3 introduced a dedicated `str` dtype by default, replacing the generic `object` dtype previously used for strings.  The notebooks worked if I used a virtual environment with Pandas 2.  With assistance from Claude Code (c.f. Disclaimer), I believe the problem is in the `formatAF()` function, where `df.allele_freq.dtype` will always be `False`. (`HLAfreq.py` line 231).  The unit tests didn't catch this, as they bypass the download path entirely.   The package should either pin `pandas<3` in `setup.py` or fix the dtype check to use `pd.api.types.is_string_dtype()`.

## Items for editorial consideration:
- [ ] **Contribution and authorship:** The submitting author David Wells is clearly identifiable throughout the commit history. Co-author Michael McAuley is not part of the commit history of the repository.  I have a memory that a previous version of the paper described McAuley's contributions to the Bayesian framework, but this is missing from the version I'm reading.  Some statement of McAuley's contribution to this work is needed.
- [ ] **Research impact statement:** A "Research Impact Statement" section is present: "HLAfreq has been used in the design of several vaccines and immunotherapies by Barinthus Biotherapeutics." This claim is plausible but unverifiable — no external publications, citations, or other specific evidence of external adoption are provided. However, given that the lead author is employed by Barinthus Biotheraputics, I do not see evidence of use beyond the authors' groups.  However, the research impacts statement is definitely more than "statements about potential future use without concrete evidence."

## Items with no issues:

- [x] **Repository**: The repository is publicly accessible.
- [x] **License**: The `LICENSE` file is an MIT license.  Note: `LICENSE` references "Vaccitech PLC", while the lead author's affiliation in the paper is "Barinthus Biotherapeutics".  Vaccitech is apparently the previous name for Barinthus Biothereaputics, so updating the `LICENSE` file may be appropriate.
- [x] **Scope and significance:** HLAfreq addresses a clear and well-defined research need: automated downloading and statistical combination of HLA allele frequency data from the Allele Frequency Net Database (AFND). HLA allele frequency is relevant to vaccine design, transplantation medicine, and autoimmune disease research. This seems appropriate to me for publication in JOSS, based my understanding of the JOSS guidelines.
- [x] **Data Sharing:** n/a
- [x] **Reproducibility:** n/a
- [x] **Human and animal research:** n/a

- [x] **Development timeline:** The first commit in the repository is in July 2022.
- [x] **Open development:** All of the commits to the repository are by the lead author.  However, there are issues created by three other individuals.  There are two forks of the repositories and 8 stars from people at diverse institutions.  While not particularly broad, this is sufficient evidence of open development from my perspective.
- [x] **Collaborative effort:** All of the commits are from the lead author, so the commit history does not show contributions from multiple developers.  However, the closed issues do provide evidence of iterative refinement through community feedback.  I do not see this as something to block publication, but that's an editorial decision.  I interpret this as "OK" for this element under the review criteria.
- [x] **Good practices:** The project has been developed over years and shows evidence of use.  There is a good amount of documentation (including Jupyter notebooks) and tests, and there is a clear contribution/support path that has been used (as evidenced by closed issues). There is not, however, evidence of a formal release process. `setup.py` has a version of "0.0.5" in line 5.  However, `__init__.py` does not have a `__version__` attribute and there are no formal releases in the repository.  I also was not able to find any sort of changelog.

- [x] **Installation:** The package installed and the tests all passed.  However, I discovered a Pandas compatibility issue (see below), where there is at least one problem if using the recently-released Pandas 3.  I always test with a fresh install into a virtual environment, which pulled the latest versions of the packages.  The simplest fix is to pin the pandas version in setup.  Apart from that, the documentation provides clear installation instructions for conda and pip (which is what I used).
- [x] **Functionality:** The supplied Jupyter notebooks provided what I consider to be a good functionality test.  Once I addressed the Pandas version issue, I was able to run `single_country.ipynb` and `working_with_priors.ipynb` with no issues.  `credible_intervals.ipynb` in the cell following the documentation for "Inappropriate prior".  That code block has `prior=[1/k]*k`, which fails because `k` is not defined.  I commented this cell out, and `credible_intervals.ipynb` ran correctly.  I ran into an network issue with `multi_country.ipynb`.  I suspect a longer timeout would be needed, but I didn't dig into that as the others ran with minimal issues.

- [x] **Statement of need:** Both the README and paper clearly state the problem (manual, slow data download from AFND; difficulty combining data across studies due to resolution differences and missing alleles), the target audience (researchers working on vaccine design, transplantation, and HLA genetics), and the motivation.
- [x] **Installation instructions:** Step-by-step instructions are provided for conda and pip, with platform-specific troubleshooting (Windows multiprocessing, g++ warnings). Dependencies are listed in `setup.py`. A `requirements.txt` is not present, but the CI installs via `pip install .`.
- [x] **Example usage:** Four Jupyter notebooks are provided in `examples/` covering single-country, multi-country, working with priors, and credible intervals. A `quickstart.py` script is also provided. Rendered HTML versions are published to GitHub Pages via CI. The README includes a minimal working example.
- [x] **Functionality documentation:** All public functions have Google-style docstrings. Documentation is generated by `pdoc` and published automatically to GitHub Pages on push to `main`. The online docs are accessible at https://barinthusbio.github.io/HLAfreq/HLAfreq.html.
- [x] **Community guidelines:** `CONTRIBUTING.md` is present and appropriate — it covers seeking support, reporting bugs, and contributing code.

- [x] **Summary:** The paper opens with a clear summary of HLAfreq's purpose, aimed at a broad audience: "a Python package which can be used to download, combine and analyze multiple HLA allele frequency datasets." The motivation for HLA research is explained (vaccine design, autoimmune disease, transplantation).
- [x] **Statement of need:** The paper identifies AFND as the primary data source, explains that current data access is "manual and slow," and lists the specific data harmonization challenges that HLAfreq addresses (incomplete studies, allele resolution differences, missing alleles).
- [x] **Software design:** The paper includes a "Software Design" section explaining the choice of Python over R (to leverage `requests` and `bs4` for AFND automated access), and the choice to return data in `pandas` DataFrames rather than custom classes ("for familiarity and in line with Scientific-Python recommendations").
- [x] **AI usage disclosure:** The paper includes an explicit disclosure: "No generative AI tools were used in the development of this software, the writing of this manuscript, or the preparation of supporting materials."
- [x] **Quality of writing:** The paper is clearly structured and readable. Mathematical exposition in the Methods section is thorough and appropriate.
  - Minor typo in "Software Design" section: `"the data are return  in"` (double space; "return" should be "returned").
- [x] **References:** References are complete and properly formatted. DOIs are provided for most entries.

## Observations for development team consideration:

- Multiple places in documentation and docstrings use an "http://" protocol descriptor.  All of these I tried  redirect to "https://".  From a security best practices perspective, these should all be changed to "https://".

- Three bibliography entries (`Gonzalez-Galarza2020`, `Weiskopf2013`, `Sidney2008`) contain `file` fields with local paths such as `file = {:home/dwells/Downloads/gkz1029.pdf:pdf}`. These are harmless to compilation but are reference-manager artifacts that should be removed.  There are also some references that aren't used, though some of these might be appropriate for a *State of the field* section.

***DISCLAIMER***: I did use Claude Code to perform an assessment of this software package and to troubleshoot the issues I found with the Jupyter Notebooks.  I reviewed each observation in that analysis to determine whether it was correct and whether it was material to my review.  All conclusions are my own and are based on observations I directly made of the repository contents.
