### 2026-03-10 #10122 @jdrugo
My review is based on commit `f753c9b` in the `main` branch, and commit `c090571` in the `paper` branch.

Overall, the paper looks good to me, other than the listed blocking issue. I agree with @usethedata that the panda 3 issues should be addressed. I have furthermore listed a few additional minor points that should be addressable by revising the paper.

### Blocking issues

- [ ] **State of the field:** The paper lacks this section. It does not describe the relationship of HLAfreq to other packages, and how HLAfreq differs from those.

### Should be addressed, but non-blocking

- [ ] **Contribution and authorship:** All commits to the repository have been made by the submitting author. The contributions of the paper's second author, Michael McAuley, should be clarified.
- [ ] **A statement of need**: The authors clearly state what problems the software is designed to solve, but the target audience is not explicitly stated.
- [ ] **Collaborative effort:** The commit history shows contributions from the submitting author. There is no public evidence for iterative refinement through community feedback.

- [ ] **Panda deprecation warning:** When installed through `conda`, the minimal example in `README.md` resulted in a pandas deprecated behavior warning in `HLAfreq.py:388` when calling `HLAfreq.decrease_resolution(aftab, 2)` and in `HLAfreq.py:550` when calling `caf = HLAfreq.combineAF(aftab)`. Pandas version 2.3.3.
- [ ] **Minimal example error:** When installed through `pip`, the minimal example in `README.md` resulted in an error in `HLAfreq.py:299` when calling `aftab = HLAfreq.only_complete(aftab)`. `pip` installs pandas v3.0.1, which makes me believe that this error is most likely related to the one identified by @usethedata.
- [ ] **References:** The list of references appears complete. Fan et al. (2017) and Salvatier, Wiecki & Fonnesbeck (2016) have DOIs that are not linked in the paper's reference section.

### No issues

- [X] **Repository:** The repository is publicly accessible.
- [X] **License:** The repository is licensed under the MIT license. The copyright is listed as 'Vaccitech PLC' which appears to be the previous name of Barinthus Biotherapeutics.
- [X] **Scope and significance:** The submission clearly states that HLAfreq supports the downloading and merging of human immune gene frequency data from a public dataset, handling various standard issues of data cleaning and wrangling.
- [X] **Data sharing:** The paper contains no original data.
- [X] **Reproducibility:** The paper contains no original results.
- [X] **Human and animal research:** The paper contains no original data.
- [X] **Development timeline:** Adequate, public commit history since July 2022.
- [X] **Open development:** The whole commit history seems to be public. All pull requests and issues were submitted/opened by the submitting author.
- [X] **Good practices:** The project contains license, clear documentation, tests, and guidelines for contributing.
- [X] **Installation:** `README.md` explains installation through `conda` and `pip`, both of which work as described.
- [X] **Functionality:** To the best of my knowledge, the software functions as intended, as verified by running the included example code and Jupyter notebooks.
- [X] **Performance:** No performance claims.
- [X] **Installation instructions:** Installation is clearly documented, and dependencies are handled either through `conda` or `pip`. Installation through `pip` causes installation of pandas >v3, which appears to lead to errors (see above).
- [X] **Example usage:** The documentation includes minimal examples, as well as examples that solve real-world analysis problems.
- [X] **Functionality documentation:** Core functionality is well-documented. API documentation is automatically generated through Github workflows, calling pdoc. The resulting documentation is publicly hosted on Github pages.
- [X] **Automated tests:** Automated tests are included, and automatically run through Github workflows. Functionality of the software can be additionally verified through running these tests locally, and by using the included example scripts.
- [X] **Community guidelines:** `CONTRIBUTING.md` is present and covers all required aspects.
- [X] **Summary:** The paper has a clear description of the high-level functionality of HLAfreq for a non-specialist audience.
- [X] **Software Design:** The paper includes this section, explaining why Python was chosen over R, and the use of pandas dataframes.
- [X] **Research Impact Statement:** The paper includes a statement on research impact. Due to the non-public nature of the research at Barinthus Biotherapeutics, the impact statement cannot be verified.
- [X] **AI usage disclosure (required section):** The paper states that no generative AI tools were used for software development or paper writing.
- [X] **Quality of writing:** The writing is clear and of acceptable quality.
