### 2026-03-17 1dd3eb1
add asserts to `test_correct_c_array` and `test_correct_c_array_alleles`
