### 2026-03-18 3357b84
check data type using is_string_dtype to handle pandas 3 update  thanks to @usethedata for diving in to this https://github.com/openjournals/joss-reviews/issues/10122#issuecomment-4020246959
