+++ examples/paper/reproduce_paper_results2.py
@@ -27,7 +27,7 @@ iedb_refa = [i for i in iedb_ref if "A" in i]
 iedb_refb = [i for i in iedb_ref if "B" in i]
 
 # Download countries in regions as defined on 
-# http://www.allelefrequencies.net/datasets.asp#tag_4
+# https://www.allelefrequencies.net/datasets.asp#tag_4
 r = requests.get("https://raw.githubusercontent.com/BarinthusBio/HLAfreq/main/data/example/countries.csv")
 with open("data/example/countries.csv", "w") as f:
     f.write(r.text)
