### 2026-03-19 d984e47
avoid depreciation warning by setting include_groups for .apply on grouped data
