### 2026-03-21 68295a5
Add author contributions section
