### 2026-03-21 518a46d
Add author contributions section
