### 2026-03-21 0c9595e
Expand research impact statement with external references
