+++ paper/paper.bib
@@ -215,3 +215,28 @@ year = {2008}
 	year = {2015},
 	pages = {101--112}
 }
+@article{niederlova2025imbalance,
+  title={Imbalance of stem-like and effector T cell states in children with early type 1 diabetes across conventional and regulatory subsets},
+  author={Niederlova, Veronika and Neuwirth, Ales and Neuman, Vit and Michalik, Juraj and Charvatova, Bela and Modrak, Martin and Sumnik, Zdenek and Stepanek, Ondrej},
+  journal={Nature Communications},
+  year={2025},
+  publisher={Nature Publishing Group UK London}
+}
+@article{testa2025scan,
+  title={SCAN-ACT: adoptive T cell therapy target discovery through single-cell transcriptomics},
+  author={Testa, Stefano and Pal, Aastha and Subramanian, Ajay and Varma, Sushama and Tang, Jack Pengfei and Graham, Danielle and Arfan, Sara and Pan, Minggui and Bui, Nam Q and Ganjoo, Kristen N and others},
+  journal={Genome Medicine},
+  volume={17},
+  number={1},
+  pages={89},
+  year={2025},
+  publisher={Springer}
+}
+@article{li2025super,
+  title={Super-pangenome analysis of 3562 human and animal papillomavirus isolates illuminates their genome and pathogenicity evolution},
+  author={Li, Huihui and Chen, Jingtao and Zhang, Xinxin and Zhang, Xing and Yang, Ludong and Ding, Yi and Chen, Congying and Shuai, Yan and Song, Ming and Liu, Jihong and others},
+  journal={bioRxiv},
+  pages={2025--07},
+  year={2025},
+  publisher={Cold Spring Harbor Laboratory}
+}
