### 2026-03-26 da10e65
Remove file references from bibliography
