+++ paper/paper.bib
@@ -103,7 +103,6 @@ MRREVIEWER = {Michael J. Evans},
 abstract = {The Allele Frequency Net Database (AFND, www.allelefrequencies.net) provides the scientific community with a freely available repository for the storage of frequency data (alleles, genes, haplotypes and genotypes) related to human leukocyte antigens (HLA), killer-cell immunoglobulin-like receptors (KIR), major histocompatibility complex Class I chain related genes (MIC) and a number of cytokine gene polymorphisms in worldwide populations. In the last five years, AFND has become more popular in terms of clinical and scientific usage, with a recent increase in genotyping data as a necessary component of Short Population Report article submissions to another scientific journal. In addition, we have developed a user-friendly desktop application for HLA and KIR genotype/population data submissions. We have also focused on classification of existing and new data into 'gold-silver-bronze' criteria, allowing users to filter and query depending on their needs. Moreover, we have also continued to expand other features, for example focussed on HLA associations with adverse drug reactions. At present, AFND contains >1600 populations from >10 million healthy individuals, making AFND a valuable resource for the analysis of some of the most polymorphic regions in the human genome.},
 author = {Gonzalez-Galarza, Faviel F. and McCabe, Antony and Santos, Eduardo J.Melo Dos and Jones, James and Takeshita, Louise and Ortega-Rivera, Nestor D. and Cid-Pavon, Glenda M.Del and Ramsbottom, Kerry and Ghattaoraya, Gurpreet and Alfirevic, Ana and Middleton, Derek and Jones, Andrew R.},
 doi = {10.1093/nar/gkz1029},
-file = {:home/dwells/Downloads/gkz1029.pdf:pdf},
 issn = {13624962},
 journal = {Nucleic Acids Research},
 number = {D1},
@@ -150,7 +149,6 @@ year = {2020}
 abstract = {The role of CD8+ T cells in dengue virus infection and subsequent disease manifestations is not fully understood.According to the original antigenic sin theory, skewing of T-cell responses induced by primary infection with one serotype causes less effective response upon secondary infection with a different serotype, predisposing individuals to severe disease. A comprehensive analysis of CD8+ responses in the general population from the Sri Lankan hyperendemic area, involving the measurement of ex vivo IFN? responses associated with more than 400 epitopes, challenges the original antigenic sin theory. Although skewing of responses toward primary infecting viruses was detected, this was not associated with impairment of responses either qualitatively or quantitatively. Furthermore, wedemonstrate highermagnitude andmore polyfunctional responses for HLA alleles associated with decreased susceptibility to severe disease, suggesting that a vigorous response by multifunctional CD8+ T cells is associated with protection from dengue virus disease.},
 author = {Weiskopf, Daniela and Angelo, Michael A. and {De Azeredo}, Elzinandes L. and Sidney, John and Greenbaum, Jason A. and Fernando, Anira N. and Broadwater, Anne and Kolla, Ravi V. and {De Silva}, Aruna D. and {De Silva}, Aravinda M. and Mattia, Kimberly A. and Doranz, Benjamin J. and Grey, Howard M. and Shresta, Sujan and Peters, Bjoern and Sette, Alessandro},
 doi = {10.1073/pnas.1305227110},
-file = {:home/dwells/Downloads/pnas.201305227.pdf:pdf},
 issn = {00278424},
 journal = {Proceedings of the National Academy of Sciences of the United States of America},
 number = {22},
@@ -175,7 +173,6 @@ year = {2013}
[…]
