### 2026-03-26 ffb2414
Add State of the field section
