### 2026-03-26 f5bc8a9
add dois for fan and salvatier
