+++ paper/paper.bib
@@ -55,6 +55,7 @@ MRREVIEWER = {Michael J. Evans},
 @article{fan2017hla,
   title={HLA association with drug-induced adverse reactions},
   author={Fan, Wen-Lang and Shiao, Meng-Shin and Hui, Rosaline Chung-Yee and Su, Shih-Chi and Wang, Chuang-Wei and Chang, Ya-Ching and Chung, Wen-Hung},
+  doi={10.1155/2017/3186328},
   journal={Journal of immunology research},
   volume={2017},
   year={2017},
@@ -116,6 +117,7 @@ year = {2020}
 @article{salvatier2016probabilistic,
   title={Probabilistic programming in Python using PyMC3},
   author={Salvatier, John and Wiecki, Thomas V and Fonnesbeck, Christopher},
+  doi={10.7717/peerj-cs.55},
   journal={PeerJ Computer Science},
   volume={2},
   pages={e55},
