### 2026-03-26 #10122 @DAWells
Thanks everybody for your work on this review so far. Following your comments I have updated the `main` branch to `2debf53`  and `paper` to `f5bc8a9`.

Links to commits addressing specific points below. Updates to the code (pandas3 compatibility, tests, deprciation warning) have been released on [pypi](https://pypi.org/project/HLAfreq/) and [bioconda](https://anaconda.org/channels/bioconda/packages/hlafreq/overview) as HLAfreq 0.0.6.

- [Add State of the field section](https://github.com/BarinthusBio/HLAfreq/commit/ffb24144281db269185889d351c78d978d5c092d)
- [check data type using is_string_dtype to handle pandas 3 update](https://github.com/BarinthusBio/HLAfreq/commit/3357b84d6a3355315738801ca51eea30c68ca9b0)
- [Expand research impact statement with external references](https://github.com/BarinthusBio/HLAfreq/commit/0c9595ee55baa0faad0b03ba2b6f00f52fd77927)
- [Add author contributions section](https://github.com/BarinthusBio/HLAfreq/commit/518a46d3c5e860120d8453946639e41b2e7ebcde)
- [add asserts to test_correct_c_array and test_correct_c_array_alleles](https://github.com/BarinthusBio/HLAfreq/commit/1dd3eb1aeb9c5425ec98f97c3dd1ef5cbb6b6c49)
- [avoid depreciation warning by setting include_groups for .apply on grouped data](https://github.com/BarinthusBio/HLAfreq/commit/d984e475180d9064dead9b86ea701f8787056fb2)
- [add dois for fan and salvatier](https://github.com/BarinthusBio/HLAfreq/commit/f5bc8a99cd9b364f35702fc46552842e89d0e886)
- [Explicitly state target audience](https://github.com/BarinthusBio/HLAfreq/commit/50ecd881524e2d0bb59977a0af9c32dae0cb1078)
- [Remove file references from bibliography](https://github.com/BarinthusBio/HLAfreq/commit/da10e651565c62804b855aa94686a45d365dcef0)
- [use https for url links](https://github.com/BarinthusBio/HLAfreq/commit/60e1c0fab3d7bbff5d5a70f8ef4229e2380afbda)
