### 2026-03-29 #10122 @jdrugo
The updated review is based on commit `2debf53` in the `main` branch, and commit `f5bc8a9` in the `paper` branch.

I would like to thank the authors for updates to the code and manuscript, which have addressed all of my previously identified issues. I will here only mention changes from [my previous review](https://github.com/openjournals/joss-reviews/issues/10122#issuecomment-4034226202).

### Blocking issues that have been addressed

- [X] **State of the field:** An adequate _State of the field_ has been added to the paper.

### Non-blocking issues that have been addressed

- [X] **Contribution and authorship:** The contributions of the second author are now described in the _Contribution and authorship_ section of the paper.
- [X] **A statement of need**: The statement of need has been updated to include an explicit statement of the target audience.
- [X] **Collaborative effort:** By [fboehm's previous advice](https://github.com/openjournals/joss-reviews/issues/10122#issuecomment-4020000595), this criterion is satisfied by posting the repository in a public setting.
- [X] **Panda deprecation warning:** When installed through `conda`, the pandas deprecated behavior warning has been resolved. Pandas version 2.3.3.
- [X] **Minimal example error:** When installed through `pip`, the pandas error has been fixed. Pandas version 3.0.1.
- [X] **References:** The missing DOI's have been added to Fan et al. (2017) and Salvatier, Wiecki & Fonnesbeck (2016).
