### 2026-03-31 #10122 @usethedata
I have reviewed the updated code and paper.  All of the significant issues I identified in my initial review have been resolved.

However, the changes uncovered a significant issue, which is that one of the 64 tests now fails:

```
FAILED tests/test_HLAfreq_pymc.py::test_correct_c_array[alleles0]
```

And there's one nit to correct:  In the paper (line 69 of `paper.md`, `data are return  in` should be `data are returned in` (correcting both the extra space and the verb participle).

I'm good with publication once these issues are addressed.

I looked this failure and the associated code over, and then asked Claude Code (Sonnet 4.6) to analyze the failure and suggest options for addressing it.   My main point is that the test fails.  I have provided what I hope is useful and correct information about the possible cause and possible remediation paths.  It's coming from an AI, so human verification is needed.

**Summary**: `test_correct_c_array[alleles0]` now fails because `_make_c_array()` returns NaN instead of 0.0 for a population that has no observations of a particular allele. This is the core use-case HLAfreq is designed to handle (missing alleles across populations), so I see it as a significant issue.

**Analysis**:   The problem appears to be in the test, not in the main code body.  It arises from how `_make_c_array` builds the numpy array in `HLAfreq_pymc.py`. When `add_unmeasured=True` (the default), `unmeasured_alleles()` is called on line 43 to fill in populations that are missing certain alleles. But the fill value for those missing rows is likely `NaN` for the allele frequency, producing `NaN` in the c column (line 48: `df["c"] = df.allele_freq * df[weights]`).

When the pivot table is built directly from the DataFrame in the test (`aftab.pivot(..., values="c")` ), pandas fills missing combinations with `NaN` — but then the test computes `c_pivot` with explicit `c` values that include 0.0 (because `allele_freq = 0` for absent alleles in the simulated data). The `groupby().apply(list)` approach on line 52 then packs whatever is in the c column into the array, including any `NaNs` that `unmeasured_alleles()` introduced.

Specifically for `alleles0 = ["A*01:01", "A*01:02", "A*01:03"]` — all at 2-field resolution — after `decrease_resolution(aftab, 2)` the allele names are unchanged, but one population apparently ends up with a `NaN` c-value for one allele after `unmeasured_alleles()` is called, where `0.0` is expected.

The test in `test_HLAfreq_pymc.py:31-37` calls `_make_c_array(aftab)` with default `format=True`, which re-runs `formatAF()` on data that already has numeric `allele_freq` and `sample_size columns`. With the `is_string_dtype()` guards added in commit `3357b84`, `formatAF()` now skips numeric conversion — but that shouldn't produce `NaN`. The more likely source is in how `unmeasured_alleles()` assigns frequency/count values for absent alleles, followed by the `c = allele_freq * 2n` multiplication producing `NaN` if either column is `NaN` for those rows.

**Options to fix**:

- Option A — apply `unmeasured_alleles` before computing `c_pivot`

Mirrors what `_make_c_array` does internally, so both sides of the assertion are built from the same filled data:

```python
def test_correct_c_array(aftab):
    aftab = HLAfreq.only_complete(aftab)
    aftab = HLAfreq.decrease_resolution(aftab, 2)
    aftab = HLAfreq.unmeasured_alleles(aftab)          # add this line
    aftab['c'] = 2 * aftab.allele_freq * aftab.sample_size
    c_pivot = aftab.pivot(columns="allele", index="population", values="c")
    c_array = HLAhdi._make_c_array(aftab)
    assert pytest.approx(c_array[0]) == c_pivot
```

- Option B — use pivot_table with fill_value=0

Replace `.pivot()` with `.pivot_table()`, which accepts an explicit fill value for missing combinations:

```python
c_pivot = aftab.pivot_table(
    columns="allele", index="population", values="c",
    aggfunc="sum", fill_value=0
    )
```
- Option C — fill `NaN` after the pivot (minimal change)

```python
c_pivot = aftab.pivot(columns="allele", index="population", values="c").fillna(0)
```
All three options produce the same numerical result. Option A is the most explicit about intent.
