### 2026-04-06 #10122 @assignUser
I have finished my review based on `main@2debf53` and `paper@f5bc8a9`.

~I did not run into the test failure described by @usethedata above when using python 3.12 and installing from the repo but I can replicate it using the conda version.~ Actually the issue seems to depend on the simulated input for the test as I did see it fail and pass both when re-running.

I agree with @usethedata  that this is the only blocking issue, once it is fixed I recommend publication!

I do however have some recommendations:

- The actions used in the CI workflows are heterogenous and in parts very out of date (e.g. `action/setup-python@v3`). This maintenance chore can be easily automated using [Dependabot](https://docs.github.com/en/code-security/tutorials/secure-your-dependencies/dependabot-quickstart-guide#enabling-dependabot-for-your-repository). I would recommend a [cooldown](https://docs.github.com/en/code-security/reference/supply-chain-security/dependabot-options-reference#cooldown) of ~2 weeks, this will give sometime for critical issues to be found and fixed, avoiding repeated updates or use of broken/dangerous versions.
- HLAfreq does not use any dynamic metadata, therefor I would recommend to move [everything into `pyproject.toml`](https://packaging.python.org/en/latest/guides/modernize-setup-py-project/#). This aligns with current recommendations in the python eco-system.
- I see you are using nox with flake8 which is great. I heartily recommend [ruff](https://docs.astral.sh/ruff), it is a combined formatter and linter and combines more or less all common lint rules into one fast and very configurable tool.
- Instead of just running tests/linters in CI something like https://prek.j178.dev/ (or the original pre-commit) can make it very easy to create git-hooks to run things prior to commits or pushes, giving much faster feedback than waiting for CI, without having to run things manually.
