### 2026-04-06 7d34504
Fix `test_correct_c_array` test
Fill missing values in pivot with zero to match make_c_array where zeros are important.
