### 2026-04-06 3124b9c
Create dependabot.yml
