### 2026-04-06 396003e
Fix typo in paper
double space and verb tense
