### 2026-04-06 #10122 @DAWells
Thank you again to everybody for their work on this review, I have now fixed the updated test and corrected the typo in the paper.
- [Fix test_correct_c_array test](https://github.com/BarinthusBio/HLAfreq/commit/7d3450453ba4f403f60cc338912a830d9b8c4ed7)
    The error was caused by missing values which are filled with 0 by `_make_c_array` but not by `pivot`, missing values are converted to 0 because this is important for calculations.
- [Fix typo in paper](https://github.com/BarinthusBio/HLAfreq/commit/396003ebbebde44ac3dc54f88bcba3a8bb84fe73)
    "return" corrected to "returned" and double space deleted.
