### 2026-04-07 #10122 @usethedata
> [@usethedata](https://github.com/usethedata) - it looks like a couple boxes remain unchecked in your review checklist. Do the authors need to address anything to satisfy those criteria?

Thanks for the reminder.  I've checked everything off and I'm good with proceeding.
