+++ .github/dependabot.yml
@@ -11,3 +11,9 @@ updates:
       interval: "weekly"
     cooldown:
       default-days: 14
+  - package-ecosystem: "github-actions" # See documentation for possible values
+    directory: "/" # Location of package manifests
+    schedule:
+      interval: "weekly"
+    cooldown:
+      default-days: 14
