### 2026-04-09 #10122 @DAWells
Thanks for the reccomendations @assignUser I have added dependabot to help keep things up to date. The other changes I will leave for now because everything is working as is but I will follow your suggestions for future projects 👍

The pymc install has been tempramental in the past so I am cautious about moving everything into pyproject.toml in case it causes issues with the conda recipe. The [most important part](https://packaging.python.org/en/latest/guides/modernize-setup-py-project/#) of the pyproject.toml is the build system definition which is already included.

I've used pre-commit before but havent gotten comffortable with it.
