### 2026-04-09 #10122 @fboehm
@DAWells  - can you double-check that the bib entries that currently lack DOIs really don't have DOIs? For example, please check the section "Missing DOIs" to see if those entries correspond to your cited works.
