### 2026-04-12 7aa0701
add missing dois for papers
