+++ paper/paper.bib
@@ -219,6 +219,7 @@ year = {2008}
   author={Niederlova, Veronika and Neuwirth, Ales and Neuman, Vit and Michalik, Juraj and Charvatova, Bela and Modrak, Martin and Sumnik, Zdenek and Stepanek, Ondrej},
   journal={Nature Communications},
   year={2025},
+  doi = {10.1038/s41467-025-66459-4},
   publisher={Nature Publishing Group UK London}
 }
 @article{testa2025scan,
@@ -229,6 +230,7 @@ year = {2008}
   number={1},
   pages={89},
   year={2025},
+  doi = {10.1186/s13073-025-01514-9},
   publisher={Springer}
 }
 @article{li2025super,
@@ -237,5 +239,6 @@ year = {2008}
   journal={bioRxiv},
   pages={2025--07},
   year={2025},
+  doi = {10.1101/2025.07.20.664904},
   publisher={Cold Spring Harbor Laboratory}
 }
