### 2026-04-13 #10122 @DAWells
Thanks @fboehm I have completed the tasks in the author check list and your requests on the paper.

- version 0.0.6 has been released https://github.com/BarinthusBio/HLAfreq/releases/tag/v0.0.6
- This release has been archived at https://doi.org/10.5281/zenodo.19559638
- the authors and affiliations are correct in the paper and match the archive
- [add missing dois for papers](https://github.com/BarinthusBio/HLAfreq/commit/7aa0701e99b997a112f382311e3a05f7a887efba)
- [replace colon with comma](https://github.com/BarinthusBio/HLAfreq/commit/2317864ecbcce64cd6fd058e15613f37c8f80c84)
