### 2026-04-15 #10122 @fboehm
I think that the references are now complete - it looks like those entries listed under Skip DOIs are books or web resources.
