### 2026-04-15 #10122 @fboehm
The DOIs in the pdf all direct to their intended targets.
