### 2026-04-21 #10122 @DAWells
The final proof looks good to me 👍
