### 2026-05-02 #10122 @arfon
@DAWells could you please fix these two minor issues:

**Missing word:**

- "researchers interested HLA frequencies" → "researchers interested in HLA frequencies" (Statement of need)

**Capitalisation:**

- python → Python (Software Design, first sentence)
