### 2026-05-02 #10122 @DAWells
@arfon good spot!

I've now fixed those issues with this commit: [Fix typos](https://github.com/BarinthusBio/HLAfreq/commit/4dfa3d10def24fe5ebb6f9533cc91723ead7030d)
