### 2026-05-03 #10122 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
