### 2026-05-03 #10122 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
